+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog (nionui)
 - Minor improvements to drawing performance.
 - Add window file path support to Window.
 - Add support for directly getting last or current keyboard modifiers.
+- Only change canvas item focus when no key modifiers are enabled.
 
 0.4.5 (2021-02-02)
 ------------------